feat(hero): implement countdown and refactor cost form
Adds a dynamic countdown timer to the Hero component and refactors the cost calculation form into atomic components.

- Implemented a dynamic countdown timer using useState and useEffect, targeting a Sept 22, 2025 end date.
- Added responsive CSS to the timer for improved layout on mobile screens (<768px).
- Refactored the  function to resolve a TypeScript error (TS2339) by ensuring a consistent return object shape.
- Restructured the multistep cost calculation form into smaller, atomic components for improved maintainability, reusability, and code structure.
