+++ src/components/Services/CostCalculation/MultiStepForm.jsx
@@ -1,54 +1,22 @@
 import React, { useState, useCallback, useRef } from "react";
-import PropTypes from "prop-types";
 import { loadStripe } from "@stripe/stripe-js";
 import {
   ThemeProvider,
-  createTheme,
   Box,
   Container,
   Typography,
   Button,
-  Grid,
-  Card,
-  CardContent,
-  Chip,
   Stepper,
   Step,
   StepLabel,
-  CircularProgress,
   Snackbar,
   Alert,
-  TextField,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  Divider,
-  Fade,
-  CardActionArea,
-  Paper,
-  Grow,
-  Zoom,
+  // Fade,
   Stack,
-  Collapse,
 } from "@mui/material";
 import {
-  ChevronRight,
-  ChevronLeft,
   Check,
-  Wifi,
-  Building,
-  Globe,
-  Rocket,
   RefreshCw,
-  Palette,
-  Megaphone,
-  Settings,
-  LocateOff as LocalOffer,
-  Bus as Business,
-  Ruler as Schedule,
-  ChevronDown,
-  ChevronUp,
 } from "lucide-react";
 
 import {
@@ -58,6 +26,17 @@ import {
   generateCampaignDuration,
 } from "./components/api.js";
 import { applyDiscount } from "../../user-dashboard/utils.js";
+import { IndustrySelection } from "./steps/IndustrySelection.jsx";
+import BrandPackageSelection from "./steps/BrandPackageSelection.jsx";
+import CampaignPackageSelection from "./steps/CampaignPackages.jsx";
+import TrinityPackageSelection from "./steps/TrinityPacakageSelection.jsx";
+import PlatformTierSelection from "./steps/PlatformTierSelection.jsx";
+import StoreTypeSelection from "./steps/StoreTypeSelection.jsx";
+import FinalSummary from "./steps/FinalSummary.jsx";
+import { theme } from "../../../theme.js";
+import { GradientText } from "../../../sections/components/common/GradientText.jsx";
+import ServiceSelection from "./steps/ServiceSelection.jsx";
+import { BRAND_PACKAGES, CAMPAIGN_PACKAGES, PLATFORM_TIERS, SERVICES, TRINITY_OPTIONS } from "./data/costCalculationData.js";
 
 const STRIPE_KEY = import.meta.env.VITE_STRIPE_PUBLIC_KEY || "";
 let stripePromise = null;
@@ -69,2543 +48,6 @@ if (STRIPE_KEY) {
   );
 }
 
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#6366F1",
-      light: "#818CF8",
-      dark: "#4F46E5",
-      contrastText: "#FFFFFF",
-    },
-    secondary: {
-      main: "#EC4899",
-      light: "#F472B6",
-      dark: "#DB2777",
-    },
-    background: {
-      default: "#F8FAFC",
-      paper: "#FFFFFF",
-    },
-    success: {
-      main: "#10B981",
-      light: "#34D399",
-      dark: "#059669",
-    },
-    warning: {
-      main: "#F59E0B",
-      light: "#FBBF24",
-      dark: "#D97706",
-    },
-    error: {
-      main: "#EF4444",
-      light: "#F87171",
-      dark: "#DC2626",
-    },
-    grey: {
-      50: "#F9FAFB",
-      100: "#F3F4F6",
-      200: "#E5E7EB",
-      300: "#D1D5DB",
-      400: "#9CA3AF",
-      500: "#6B7280",
-      600: "#4B5563",
-      700: "#374151",
-      800: "#1F2937",
-      900: "#111827",
-    },
-    text: {
-      primary: "#111827",
-      secondary: "#6B7280",
-    },
-  },
-  typography: {
-    fontFamily: '"Inter", "Roboto", "Helvetica", "Arial", sans-serif',
-    h1: {
-      fontSize: "3.5rem",
-      fontWeight: 800,
-      lineHeight: 1.1,
-      letterSpacing: "-0.025em",
-      "@media (max-width:600px)": {
-        fontSize: "2.5rem",
-      },
-    },
-    h2: {
-      fontSize: "2.25rem",
-      fontWeight: 700,
-      lineHeight: 1.2,
-      letterSpacing: "-0.025em",
-      "@media (max-width:600px)": {
-        fontSize: "1.875rem",
-      },
-    },
-    h3: {
-      fontSize: "1.875rem",
-      fontWeight: 600,
-      lineHeight: 1.3,
-    },
-    h4: {
-      fontSize: "1.5rem",
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    h5: {
-      fontSize: "1.25rem",
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    h6: {
-      fontSize: "1.125rem",
-      fontWeight: 600,
-      lineHeight: 1.4,
-    },
-    body1: {
-      fontSize: "1rem",
-      lineHeight: 1.6,
-    },
-    body2: {
-      fontSize: "0.875rem",
-      lineHeight: 1.6,
-    },
-    button: {
-      fontWeight: 600,
-      textTransform: "none",
-    },
-  },
-  shape: {
-    borderRadius: 16,
-  },
-  components: {
-    MuiButton: {
-      styleOverrides: {
-        root: {
-          borderRadius: 12,
-          padding: "12px 24px",
-          fontSize: "0.95rem",
-          fontWeight: 600,
-          boxShadow: "none",
-          "&:hover": {
-            boxShadow: "0 4px 12px rgba(99, 102, 241, 0.3)",
-            transform: "translateY(-1px)",
-          },
-        },
-        contained: {
-          background: "linear-gradient(135deg, #6366F1 0%, #8B5CF6 100%)",
-          "&:hover": {
-            background: "linear-gradient(135deg, #4F46E5 0%, #7C3AED 100%)",
-          },
-        },
-      },
-    },
-    MuiCard: {
-      styleOverrides: {
-        root: {
-          borderRadius: 20,
-          border: "1px solid rgba(0, 0, 0, 0.05)",
-          transition: "all 0.3s cubic-bezier(0.4, 0, 0.2, 1)",
-        },
-      },
-    },
-    MuiChip: {
-      styleOverrides: {
-        root: {
-          borderRadius: 12,
-          fontWeight: 600,
-        },
-      },
-    },
-    MuiTextField: {
-      styleOverrides: {
-        root: {
-          "& .MuiOutlinedInput-root": {
-            borderRadius: 12,
-          },
-        },
-      },
-    },
-    MuiStepper: {
-      styleOverrides: {
-        root: {
-          background: "transparent",
-        },
-      },
-    },
-  },
-});
-
-const SERVICES = [
-  {
-    id: "brand-identity",
-    name: "Brand Identity Package",
-    icon: Palette,
-    description:
-      "Professional brand identity design and strategy that sets you apart",
-    color: "#E11D48",
-    gradient: "linear-gradient(135deg, #E11D48 0%, #F97316 100%)",
-    benefits: ["Logo Design", "Brand Guidelines", "Marketing Materials"],
-  },
-  {
-    id: "ad-campaigns",
-    name: "Digital Ad Campaigns",
-    icon: Megaphone,
-    description: "Targeted digital marketing campaigns that drive real results",
-    color: "#059669",
-    gradient: "linear-gradient(135deg, #059669 0%, #10B981 100%)",
-    benefits: ["Google Ads", "Meta Ads", "Performance Analytics"],
-  },
-  {
-    id: "trinity-systems",
-    name: "Trinity Business Systems",
-    icon: Settings,
-    description: "AI-powered business intelligence platforms for growth",
-    color: "#7C3AED",
-    gradient: "linear-gradient(135deg, #7C3AED 0%, #6366F1 100%)",
-    benefits: ["AI Optimization", "Predictive Analytics", "Automation"],
-  },
-  {
-    id: "complete-platform",
-    name: "Complete Digital Platform",
-    icon: Rocket,
-    description: "Full-service website and business systems solution",
-    color: "#DC2626",
-    gradient: "linear-gradient(135deg, #DC2626 0%, #EA580C 100%)",
-    benefits: ["Custom Website", "E-commerce", "Full Integration"],
-  },
-];
-
-const BRAND_PACKAGES = [
-  {
-    id: "starter",
-    name: "Starter Brand Identity",
-    price: 500,
-    popular: false,
-    description: "Perfect for new businesses starting their brand journey",
-    features: [
-      "Logo design concept (up to 2 proposals)",
-      "Colour scheme development",
-      "Typography selection",
-      "Business card design",
-      "Letterhead or invoice design",
-      "Brand questionnaire & style scope",
-      "1 revision round",
-    ],
-    deliveryTime: "5-7 business days",
-  },
-  {
-    id: "premium",
-    name: "Premium Brand Identity",
-    price: 750,
-    popular: true,
-    description: "Most popular choice for established businesses",
-    features: [
-      "Logo design concept (up to 2 proposals)",
-      "Comprehensive colour scheme",
-      "Typography system",
-      "Business card design",
-      "Letterhead or invoice design",
-      "Brand questionnaire & style scope",
-      "Brand dos and don'ts guide",
-      "Up to 3 marketing materials",
-      "2 revision rounds",
-    ],
-    deliveryTime: "7-10 business days",
-  },
-  {
-    id: "elite",
-    name: "Elite Brand Identity",
-    price: 1000,
-    popular: false,
-    description: "Complete brand system for premium businesses",
-    features: [
-      "Logo design concept (up to 2 proposals)",
-      "Full colour system",
-      "Typography hierarchy",
-      "Complete stationery suite",
-      "Brand questionnaire & style scope",
-      "Comprehensive brand guidelines",
-      "Up to 5 marketing materials",
-      "4 revision rounds",
-      "Brand personality overview",
-      "Printing & digital asset guides",
-      "1-hour consultation meeting",
-    ],
-    deliveryTime: "10-14 business days",
-  },
-];
-
-const CAMPAIGN_PACKAGES = [
-  {
-    id: "starter",
-    name: "Starter Campaign",
-    price: 350,
-    description: "Perfect for small businesses starting digital marketing",
-    icon: "🚀",
-    features: [
-      "Basic customer segmentation",
-      "Competitor scan + keyword research (10-15 keywords)",
-      "1x Google Ads campaign setup",
-      "Up to 4 custom ad designs",
-      "Performance report after 2 months",
-    ],
-    deliveryTime: "3-5 business days",
-  },
-  {
-    id: "growth",
-    name: "Growth Campaign",
-    price: 500,
-    description: "Most popular for SMEs seeking wider reach",
-    featured: true,
-    icon: "📈",
-    features: [
-      "Advanced customer segmentation",
-      "Comprehensive competitor analysis",
-      "2x Google Ads campaigns with optimization",
-      "Up to 7 premium ad designs",
-      "Mid-campaign update + final report",
-      "CTR, conversions & ROI insights",
-    ],
-    deliveryTime: "5-7 business days",
-  },
-  {
-    id: "premium",
-    name: "Premium Campaign",
-    price: 750,
-    description: "Complete solution for market domination",
-    icon: "🏆",
-    features: [
-      "Deluxe segmentation + 2 buyer personas",
-      "In-depth competitor benchmarking",
-      "3x Google Ads campaigns with ongoing optimization",
-      "Up to 10 premium ad designs with copywriting",
-      "Detailed analytics dashboard",
-      "Strategic recommendations & next steps",
-    ],
-    deliveryTime: "7-10 business days",
-  },
-];
-
-const TRINITY_OPTIONS = [
-  {
-    id: "expense",
-    type: "individual",
-    name: "Expense Manager",
-    icon: "💰",
-    description:
-      "AI-powered financial planning that ensures you never run out of cash",
-    betaPrice: 230,
-    standardPrice: 2074,
-    savings: 87,
-    features: [
-      "Predictive cash flow modeling",
-      "7 intelligent allocation rules",
-      "Emergency fund protection",
-      "Automatic priority reshuffling",
-      "Real-time expense tracking",
-    ],
-    benefits: [
-      "Avoid cash shortages",
-      "Optimize spending",
-      "Build emergency funds",
-    ],
-  },
-  {
-    id: "mcd",
-    type: "individual",
-    name: "MCD System",
-    icon: "📈",
-    description:
-      "Marketing Cost Displacement - automatically adjusts prices based on ad spend",
-    betaPrice: 200,
-    standardPrice: 1500,
-    savings: 87,
-    features: [
-      "Real-time price optimization",
-      "Marketing spend integration",
-      "Automatic profit protection",
-      "Platform cost tracking",
-      "Dynamic pricing algorithms",
-    ],
-    benefits: [
-      "Maintain profit margins",
-      "Optimize ad ROI",
-      "Automated pricing",
-    ],
-  },
-  {
-    id: "rcd",
-    type: "individual",
-    name: "RCD System",
-    icon: "🎯",
-    description:
-      "Returning Customer Discounts - creates viral loyalty networks",
-    betaPrice: 200,
-    standardPrice: 1500,
-    savings: 87,
-    features: [
-      "Automatic loyalty tracking",
-      "Viral referral networks",
-      "Personalized discount vectors",
-      "Lifetime value optimization",
-      "Customer journey mapping",
-    ],
-    benefits: ["Increase customer retention", "Viral growth", "Higher LTV"],
-  },
-  {
-    id: "garo",
-    type: "individual",
-    name: "GARO System",
-    icon: "🧬",
-    description:
-      "Genetic Algorithm Restocking Optimizer - evolves perfect inventory decisions",
-    betaPrice: 2230,
-    standardPrice: 2074,
-    savings: 87,
-    features: [
-      "500+ generation evolution",
-      "Multi-objective fitness scoring",
-      "70% stockout reduction",
-      "Predictive demand modeling",
-      "Automated reorder points",
-    ],
-    benefits: ["Reduce stockouts", "Optimize inventory", "Predict demand"],
-    note: "Physical stores: includes €1,600 for Square setup & training",
-  },
-  {
-    id: "aed",
-    type: "individual",
-    name: "AED System",
-    icon: "🚀",
-    description:
-      "Advertising Efficiency Dashboard - unifies and optimizes all ad platforms",
-    betaPrice: 230,
-    standardPrice: 1500,
-    savings: 87,
-    features: [
-      "4+ platform integration",
-      "Real-time budget reallocation",
-      "Machine learning optimization",
-      "30% cost reduction average",
-      "Cross-platform analytics",
-    ],
-    benefits: ["Unified ad management", "Reduce ad costs", "ML optimization"],
-  },
-  {
-    id: "trinity-core",
-    type: "package",
-    name: "Trinity Core Package",
-    icon: "⚡",
-    description: "Essential business intelligence: Expense Manager + MCD + RCD",
-    betaPrice: 690,
-    standardPrice: 5187,
-    includes: ["💰 Expense Manager", "📈 MCD System", "🎯 RCD System"],
-    savings: 1728,
-    baseRecommended: true,
-    benefits: [
-      "Complete financial control",
-      "Automated pricing",
-      "Customer loyalty",
-    ],
-  },
-  {
-    id: "trinity-plus",
-    type: "package",
-    name: "Trinity Plus Package",
-    icon: "🌟",
-    description: "Complete suite: All 5 systems working in perfect harmony",
-    betaPrice: 2994,
-    standardPrice: 10489,
-    includes: [
-      "💰 Expense Manager",
-      "📈 MCD System",
-      "🎯 RCD System",
-      "🧬 GARO System",
-      "🚀 AED System",
-    ],
-    savings: 4724,
-    bestValue: true,
-    benefits: ["Complete automation", "Maximum savings", "Full integration"],
-    note: "GARO requires €1,600 for physical stores",
-  },
-];
-
-const PLATFORM_TIERS = [
-  {
-    id: "foundation",
-    name: "Foundation",
-    minPrice: 600,
-    maxPrice: 1000,
-    description: "Perfect starting point for your digital presence",
-    icon: "🏗️",
-    features: [
-      "Professional responsive website",
-      "Mobile-first design",
-      "Basic SEO optimization",
-      "Contact forms & lead capture",
-      "Social media integration",
-      "SSL certificate & security",
-    ],
-    deliveryTime: "7-10 business days",
-  },
-  {
-    id: "full",
-    name: "Full System",
-    minPrice: 1200,
-    maxPrice: 2000,
-    description: "Complete business solution with advanced features",
-    icon: "🎯",
-    popular: true,
-    features: [
-      "Everything in Foundation",
-      "Complete ordering system",
-      "Customer database & CRM",
-      "Email automation workflows",
-      "Payment processing integration",
-      "Analytics & reporting dashboard",
-      "Inventory management",
-    ],
-    deliveryTime: "14-21 business days",
-  },
-  {
-    id: "premium",
-    name: "Premium Enterprise",
-    minPrice: 1800,
-    maxPrice: 3000,
-    description: "Enterprise-grade solution for scaling businesses",
-    icon: "🚀",
-    features: [
-      "Everything in Full System",
-      "Advanced analytics & insights",
-      "Multi-location support",
-      "API access & integrations",
-      "AI-powered predictions",
-      "Priority support & training",
-      "Custom integrations",
-    ],
-    deliveryTime: "21-28 business days",
-  },
-];
-
-const industries = [
-  {
-    id: "restaurant",
-    name: "Restaurant & Food",
-    icon: "🍽️",
-    examples: "Restaurants, Takeaways, Cafes, Food Delivery",
-    color: "#EF4444",
-  },
-  {
-    id: "retail",
-    name: "Retail & E-commerce",
-    icon: "🛍️",
-    examples: "Online Stores, Boutiques, Fashion, Electronics",
-    color: "#8B5CF6",
-  },
-  {
-    id: "services",
-    name: "Professional Services",
-    icon: "💼",
-    examples: "Consulting, Legal, Financial, Healthcare",
-    color: "#059669",
-  },
-  {
-    id: "hospitality",
-    name: "Hospitality & Events",
-    icon: "🏨",
-    examples: "Hotels, Venues, Tourism, Event Planning",
-    color: "#DC2626",
-  },
-  {
-    id: "property",
-    name: "Property & Real Estate",
-    icon: "🏠",
-    examples: "Agents, Property Management, Development",
-    color: "#0891B2",
-  },
-  {
-    id: "other",
-    name: "Other Industry",
-    icon: "🚀",
-    examples: "Tell us about your unique business",
-    color: "#6366F1",
-  },
-];
-
-const gradientText = {
-  background: "linear-gradient(135deg, #6366F1 0%, #8B5CF6 50%, #EC4899 100%)",
-  WebkitBackgroundClip: "text",
-  WebkitTextFillColor: "transparent",
-  backgroundClip: "text",
-};
-
-const SelectableCard = ({ children, selected, onClick, sx, hover = true }) => (
-  <Grow in timeout={500}>
-    <Card
-      elevation={0}
-      sx={{
-        cursor: "pointer",
-        height: "100%",
-        border: "2px solid",
-        borderColor: selected ? "primary.main" : "grey.200",
-        transform: selected ? "scale(1.02)" : "none",
-        transition: "all 0.3s cubic-bezier(0.4, 0, 0.2, 1)",
-        position: "relative",
-        overflow: "visible",
-        background: selected
-          ? "linear-gradient(135deg, rgba(99, 102, 241, 0.05) 0%, rgba(139, 92, 246, 0.05) 100%)"
-          : "white",
-        "&::before": selected
-          ? {
-              content: '""',
-              position: "absolute",
-              top: -2,
-              left: -2,
-              right: -2,
-              bottom: -2,
-              background: "linear-gradient(135deg, #6366F1, #8B5CF6, #EC4899)",
-              borderRadius: "inherit",
-              zIndex: -1,
-            }
-          : {},
-        "&:hover": hover
-          ? {
-              transform: "scale(1.02)",
-              borderColor: selected ? "primary.main" : "primary.light",
-              boxShadow: "0 8px 25px rgba(99, 102, 241, 0.15)",
-              "& .card-icon": {
-                transform: "scale(1.1)",
-              },
-            }
-          : {},
-        ...sx,
-      }}
-    >
-      <CardActionArea
-        onClick={onClick}
-        sx={{
-          height: "100%",
-          borderRadius: "inherit",
-          "&:hover .MuiCardActionArea-focusHighlight": {
-            opacity: 0,
-          },
-        }}
-      >
-        <CardContent
-          sx={{
-            display: "flex",
-            flexDirection: "column",
-            height: "100%",
-            p: 3,
-            position: "relative",
-          }}
-        >
-          {selected && (
-            <Zoom in>
-              <Box
-                sx={{
-                  position: "absolute",
-                  top: 16,
-                  right: 16,
-                  width: 32,
-                  height: 32,
-                  borderRadius: "50%",
-                  background: "linear-gradient(135deg, #6366F1, #8B5CF6)",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  boxShadow: "0 4px 12px rgba(99, 102, 241, 0.4)",
-                }}
-              >
-                <Check size={16} color="white" />
-              </Box>
-            </Zoom>
-          )}
-          {children}
-        </CardContent>
-      </CardActionArea>
-    </Card>
-  </Grow>
-);
-
-SelectableCard.propTypes = {
-  children: PropTypes.node.isRequired,
-  selected: PropTypes.bool,
-  onClick: PropTypes.func,
-  sx: PropTypes.object,
-  hover: PropTypes.bool,
-};
-
-const ServiceSelection = ({
-  selectedServices,
-  setSelectedServices,
-  nextStep,
-}) => {
-  const toggleService = (serviceId) => {
-    setSelectedServices((prev) =>
-      prev.includes(serviceId)
-        ? prev.filter((id) => id !== serviceId)
-        : [...prev, serviceId]
-    );
-  };
-
-  return (
-    <Fade in timeout={700}>
-      <Box sx={{ px: { xs: 2, md: 0 } }}>
-        <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-          <Chip
-            label="Step 1 of 4: Service Selection"
-            color="primary"
-            variant="outlined"
-            sx={{
-              px: 3,
-              py: 1,
-              fontSize: "0.9rem",
-              fontWeight: 600,
-            }}
-          />
-          <Typography variant="h2" gutterBottom>
-            What services do you need?
-          </Typography>
-          <Typography
-            variant="body1"
-            color="text.secondary"
-            sx={{ maxWidth: 600, fontSize: "1.1rem", lineHeight: 1.7 }}
-          >
-            Select all the services that fit your business needs. You can choose
-            multiple services for a comprehensive solution.
-          </Typography>
-
-          {selectedServices.length > 0 && (
-            <Zoom in>
-              <Chip
-                label={`${selectedServices.length} service${
-                  selectedServices.length > 1 ? "s" : ""
-                } selected`}
-                color="success"
-                variant="filled"
-                icon={<Check size={16} />}
-                sx={{
-                  py: 2,
-                  px: 3,
-                  fontSize: "0.9rem",
-                  fontWeight: 600,
-                }}
-              />
-            </Zoom>
-          )}
-        </Stack>
-
-        <Grid container spacing={4} sx={{ mb: 6 }}>
-          {SERVICES.map((service, index) => (
-            <Grid item xs={12} md={6} key={service.id}>
-              <Grow in timeout={500 + index * 100}>
-                <div>
-                  <SelectableCard
-                    selected={selectedServices.includes(service.id)}
-                    onClick={() => toggleService(service.id)}
-                  >
-                    <Box
-                      className="card-icon"
-                      sx={{
-                        width: 80,
-                        height: 80,
-                        borderRadius: 3,
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                        mb: 3,
-                        background: service.gradient,
-                        transition: "transform 0.3s ease",
-                        boxShadow: `0 8px 25px ${service.color}25`,
-                      }}
-                    >
-                      <service.icon
-                        style={{
-                          color: "white",
-                          width: 36,
-                          height: 36,
-                        }}
-                      />
-                    </Box>
-                    <Typography
-                      variant="h5"
-                      component="h3"
-                      gutterBottom
-                      fontWeight={700}
-                    >
-                      {service.name}
-                    </Typography>
-                    <Typography
-                      variant="body1"
-                      color="text.secondary"
-                      sx={{ mb: 3, lineHeight: 1.6 }}
-                    >
-                      {service.description}
-                    </Typography>
-
-                    <Stack
-                      direction="row"
-                      spacing={1}
-                      flexWrap="wrap"
-                      useFlexGap
-                    >
-                      {service.benefits.map((benefit) => (
-                        <Chip
-                          key={benefit}
-                          label={benefit}
-                          size="small"
-                          sx={{
-                            bgcolor: "grey.100",
-                            fontSize: "0.75rem",
-                            height: 24,
-                          }}
-                        />
-                      ))}
-                    </Stack>
-                  </SelectableCard>
-                </div>
-              </Grow>
-            </Grid>
-          ))}
-        </Grid>
-
-        <Box textAlign="center">
-          <Button
-            variant="contained"
-            size="large"
-            onClick={nextStep}
-            disabled={selectedServices.length === 0}
-            endIcon={<ChevronRight size={20} />}
-            sx={{
-              minWidth: 220,
-              py: 1.5,
-              fontSize: "1rem",
-              background:
-                selectedServices.length > 0
-                  ? "linear-gradient(135deg, #6366F1 0%, #8B5CF6 100%)"
-                  : undefined,
-            }}
-          >
-            Continue with {selectedServices.length} service
-            {selectedServices.length !== 1 ? "s" : ""}
-          </Button>
-        </Box>
-      </Box>
-    </Fade>
-  );
-};
-
-ServiceSelection.propTypes = {
-  selectedServices: PropTypes.array.isRequired,
-  setSelectedServices: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-};
-
-const IndustrySelection = ({
-  selectedIndustry,
-  setSelectedIndustry,
-  nextStep,
-  prevStep,
-}) => (
-  <Fade in timeout={500}>
-    <Box>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label="Step 2 of 4: Industry Selection"
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          What industry is your business in?
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 600, fontSize: "1.1rem", lineHeight: 1.7 }}
-        >
-          This helps us tailor our recommendations and customize the solution
-          for your specific industry needs.
-        </Typography>
-      </Stack>
-
-      <Grid container spacing={3} sx={{ mb: 6 }}>
-        {industries.map((industry, index) => (
-          <Grid item xs={12} sm={6} md={4} key={industry.id}>
-            <Grow in timeout={300 + index * 100}>
-              <div>
-                <SelectableCard
-                  selected={selectedIndustry === industry.id}
-                  onClick={() => setSelectedIndustry(industry.id)}
-                >
-                  <Box
-                    sx={{
-                      width: 72,
-                      height: 72,
-                      borderRadius: "50%",
-                      display: "flex",
-                      alignItems: "center",
-                      justifyContent: "center",
-                      mb: 3,
-                      mx: "auto",
-                      background: `linear-gradient(135deg, ${industry.color}20, ${industry.color}10)`,
-                      border: `2px solid ${industry.color}30`,
-                    }}
-                  >
-                    <Typography
-                      variant="h3"
-                      component="p"
-                      sx={{ fontSize: "2.5rem" }}
-                    >
-                      {industry.icon}
-                    </Typography>
-                  </Box>
-                  <Typography
-                    variant="h6"
-                    component="h3"
-                    gutterBottom
-                    textAlign="center"
-                    fontWeight={600}
-                  >
-                    {industry.name}
-                  </Typography>
-                  <Typography
-                    variant="body2"
-                    color="text.secondary"
-                    textAlign="center"
-                    sx={{ lineHeight: 1.6 }}
-                  >
-                    {industry.examples}
-                  </Typography>
-                </SelectableCard>
-              </div>
-            </Grow>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Stack direction="row" spacing={2} justifyContent="center">
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{ minWidth: 120 }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={!selectedIndustry}
-          endIcon={<ChevronRight size={20} />}
-          sx={{ minWidth: 160 }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-IndustrySelection.propTypes = {
-  selectedIndustry: PropTypes.string,
-  setSelectedIndustry: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-};
-
-const BrandPackageSelection = ({
-  selectedPackage,
-  setSelectedPackage,
-  nextStep,
-  prevStep,
-  currentServiceIndex,
-  totalServices,
-}) => (
-  <Fade in timeout={500}>
-    <Box>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label={`Service ${
-            currentServiceIndex + 1
-          } of ${totalServices}: Brand Identity`}
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          Choose Your Brand Identity Package
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 700, fontSize: "1.1rem", lineHeight: 1.7 }}
-        >
-          Select the package that best fits your brand identity needs. All
-          packages include professional design and brand strategy.
-        </Typography>
-      </Stack>
-
-      <Grid container spacing={4} sx={{ mb: 6 }}>
-        {BRAND_PACKAGES.map((pkg, index) => (
-          <Grid item xs={12} md={4} key={pkg.id}>
-            <Grow in timeout={300 + index * 150}>
-              <div>
-                <SelectableCard
-                  selected={selectedPackage === pkg.id}
-                  onClick={() => setSelectedPackage(pkg.id)}
-                  sx={{
-                    height: "100%",
-                    position: "relative",
-                  }}
-                >
-                  {pkg.popular && (
-                    <Chip
-                      label="Most Popular"
-                      color="primary"
-                      size="small"
-                      sx={{
-                        position: "absolute",
-                        top: -10,
-                        left: "50%",
-                        transform: "translateX(-50%)",
-                        zIndex: 1,
-                        bgcolor: "primary.main",
-                        color: "white",
-                        fontWeight: 600,
-                      }}
-                    />
-                  )}
-
-                  <Stack spacing={2} height="100%">
-                    <Box textAlign="center" pt={pkg.popular ? 2 : 0}>
-                      <Typography
-                        variant="h5"
-                        component="h3"
-                        gutterBottom
-                        fontWeight={700}
-                      >
-                        {pkg.name}
-                      </Typography>
-                      <Typography
-                        variant="body2"
-                        color="text.secondary"
-                        sx={{ mb: 2, lineHeight: 1.6 }}
-                      >
-                        {pkg.description}
-                      </Typography>
-                      <Box sx={{ mb: 2 }}>
-                        <Typography
-                          variant="h3"
-                          component="p"
-                          color="primary"
-                          sx={{ fontWeight: 800, mb: 0.5 }}
-                        >
-                          €{pkg.price.toLocaleString()}
-                        </Typography>
-                        <Chip
-                          label={pkg.deliveryTime}
-                          size="small"
-                          icon={<Schedule size={14} />}
-                          sx={{ bgcolor: "grey.100", fontSize: "0.75rem" }}
-                        />
-                      </Box>
-                    </Box>
-
-                    <List dense sx={{ p: 0, flexGrow: 1 }}>
-                      {pkg.features.map((feature, idx) => (
-                        <ListItem key={idx} disableGutters sx={{ py: 0.5 }}>
-                          <ListItemIcon sx={{ minWidth: 32 }}>
-                            <Box
-                              sx={{
-                                width: 20,
-                                height: 20,
-                                borderRadius: "50%",
-                                bgcolor: "success.main",
-                                display: "flex",
-                                alignItems: "center",
-                                justifyContent: "center",
-                              }}
-                            >
-                              <Check size={12} color="white" />
-                            </Box>
-                          </ListItemIcon>
-                          <ListItemText
-                            primary={feature}
-                            primaryTypographyProps={{
-                              variant: "body2",
-                              sx: { lineHeight: 1.6 },
-                            }}
-                          />
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Stack>
-                </SelectableCard>
-              </div>
-            </Grow>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Stack direction="row" spacing={2} justifyContent="center">
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{ minWidth: 120 }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={!selectedPackage}
-          endIcon={<ChevronRight size={20} />}
-          sx={{ minWidth: 160 }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-BrandPackageSelection.propTypes = {
-  selectedPackage: PropTypes.string,
-  setSelectedPackage: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-  currentServiceIndex: PropTypes.number,
-  totalServices: PropTypes.number,
-};
-
-const CampaignPackageSelection = ({
-  selectedCampaign,
-  setSelectedCampaign,
-  nextStep,
-  prevStep,
-  currentServiceIndex,
-  totalServices,
-}) => (
-  <Fade in timeout={500}>
-    <Box>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label={`Service ${
-            currentServiceIndex + 1
-          } of ${totalServices}: Digital Campaigns`}
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          Choose Your Campaign Package
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 700, fontSize: "1.1rem", lineHeight: 1.7 }}
-        >
-          Select the campaign package that matches your marketing goals and
-          budget. All packages include professional setup and optimization.
-        </Typography>
-      </Stack>
-
-      <Grid container spacing={4} sx={{ mb: 6 }}>
-        {CAMPAIGN_PACKAGES.map((campaign, index) => (
-          <Grid item xs={12} md={4} key={campaign.id}>
-            <Grow in timeout={300 + index * 150}>
-              <div>
-                <SelectableCard
-                  selected={selectedCampaign === campaign.id}
-                  onClick={() => setSelectedCampaign(campaign.id)}
-                  sx={{ height: "100%" }}
-                >
-                  {campaign.featured && (
-                    <Chip
-                      label="Most Popular"
-                      color="primary"
-                      size="small"
-                      sx={{
-                        position: "absolute",
-                        top: -10,
-                        left: "50%",
-                        transform: "translateX(-50%)",
-                        zIndex: 1,
-                        bgcolor: "primary.main",
-                        color: "white",
-                        fontWeight: 600,
-                      }}
-                    />
-                  )}
-
-                  <Stack spacing={2} height="100%">
-                    <Box textAlign="center" pt={campaign.featured ? 2 : 0}>
-                      <Typography variant="h3" sx={{ mb: 1, fontSize: "2rem" }}>
-                        {campaign.icon}
-                      </Typography>
-                      <Typography
-                        variant="h5"
-                        component="h3"
-                        gutterBottom
-                        fontWeight={700}
-                      >
-                        {campaign.name}
-                      </Typography>
-                      <Typography
-                        variant="body2"
-                        color="text.secondary"
-                        sx={{ mb: 2, lineHeight: 1.6 }}
-                      >
-                        {campaign.description}
-                      </Typography>
-                      <Box sx={{ mb: 2 }}>
-                        <Typography
-                          variant="h3"
-                          component="p"
-                          color="primary"
-                          sx={{ fontWeight: 800, mb: 0.5 }}
-                        >
-                          €{campaign.price.toLocaleString()}
-                        </Typography>
-                        <Chip
-                          label={campaign.deliveryTime}
-                          size="small"
-                          icon={<Schedule size={14} />}
-                          sx={{ bgcolor: "grey.100", fontSize: "0.75rem" }}
-                        />
-                      </Box>
-                    </Box>
-
-                    <List dense sx={{ p: 0, flexGrow: 1 }}>
-                      {campaign.features.map((feature, idx) => (
-                        <ListItem key={idx} disableGutters sx={{ py: 0.5 }}>
-                          <ListItemIcon sx={{ minWidth: 32 }}>
-                            <Box
-                              sx={{
-                                width: 20,
-                                height: 20,
-                                borderRadius: "50%",
-                                bgcolor: "success.main",
-                                display: "flex",
-                                alignItems: "center",
-                                justifyContent: "center",
-                              }}
-                            >
-                              <Check size={12} color="white" />
-                            </Box>
-                          </ListItemIcon>
-                          <ListItemText
-                            primary={feature}
-                            primaryTypographyProps={{
-                              variant: "body2",
-                              sx: { lineHeight: 1.6 },
-                            }}
-                          />
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Stack>
-                </SelectableCard>
-              </div>
-            </Grow>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Stack direction="row" spacing={2} justifyContent="center">
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{ minWidth: 120 }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={!selectedCampaign}
-          endIcon={<ChevronRight size={20} />}
-          sx={{ minWidth: 160 }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-CampaignPackageSelection.propTypes = {
-  selectedCampaign: PropTypes.string,
-  setSelectedCampaign: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-  currentServiceIndex: PropTypes.number,
-  totalServices: PropTypes.number,
-};
-
-const TrinityPackageCard = ({ option, selected, onClick }) => {
-  const [expanded, setExpanded] = useState(false);
-
-  return (
-    <SelectableCard
-      selected={selected}
-      onClick={onClick}
-      sx={{
-        height: "100%",
-        "& .MuiCardContent-root": {
-          px: { xs: 2, sm: 3 },
-          py: { xs: 2, sm: 3 },
-        },
-      }}
-    >
-      <Stack spacing={2} height="100%">
-        <Stack
-          direction="row"
-          spacing={1}
-          flexWrap="wrap"
-          useFlexGap
-          mb={1}
-          sx={{
-            "& .MuiChip-root": {
-              fontSize: { xs: "0.7rem", sm: "0.75rem" },
-            },
-          }}
-        >
-          {option.baseRecommended && (
-            <Chip
-              label="Recommended"
-              size="small"
-              sx={{
-                bgcolor: "info.main",
-                color: "white",
-                fontWeight: 600,
-              }}
-            />
-          )}
-          {option.bestValue && (
-            <Chip
-              label="Best Value"
-              size="small"
-              color="success"
-              variant="filled"
-              sx={{ fontWeight: 600 }}
-            />
-          )}
-          <Chip
-            label={`Save ${option.savings}%`}
-            size="small"
-            sx={{
-              bgcolor: "error.light",
-              color: "error.contrastText",
-              fontWeight: 600,
-            }}
-          />
-        </Stack>
-        <Stack
-          direction={{ xs: "column", sm: "row" }}
-          spacing={2}
-          alignItems={{ xs: "flex-start", sm: "flex-start" }}
-        >
-          <Box flexShrink={0}>
-            <Typography
-              variant="h3"
-              sx={{
-                fontSize: { xs: "2rem", sm: "2.5rem" },
-                textAlign: { xs: "center", sm: "left" },
-                width: { xs: "100%", sm: "auto" },
-              }}
-            >
-              {option.icon}
-            </Typography>
-          </Box>
-          <Stack spacing={1} flex={1} minWidth={0}>
-            <Stack
-              direction={{ xs: "column", sm: "row" }}
-              justifyContent="space-between"
-              alignItems={{ xs: "flex-start", sm: "flex-start" }}
-              spacing={1}
-            >
-              <Box flex={1} minWidth={0}>
-                <Typography
-                  variant="h6"
-                  component="h3"
-                  fontWeight={700}
-                  gutterBottom
-                  sx={{
-                    fontSize: { xs: "1.1rem", sm: "1.25rem" },
-                    lineHeight: 1.2,
-                    wordBreak: "break-word",
-                    hyphens: "auto",
-                  }}
-                >
-                  {option.name}
-                </Typography>
-              </Box>
-              <Box
-                textAlign={{ xs: "left", sm: "right" }}
-                flexShrink={0}
-                width={{ xs: "100%", sm: "auto" }}
-              >
-                <Typography
-                  variant="h5"
-                  sx={{
-                    color: "error.main",
-                    fontWeight: 800,
-                    mb: 0.5,
-                    fontSize: { xs: "1.5rem", sm: "1.875rem" },
-                  }}
-                >
-                  €{option.betaPrice.toLocaleString()}
-                </Typography>
-                <Typography
-                  variant="body2"
-                  sx={{
-                    textDecoration: "line-through",
-                    color: "text.disabled",
-                    fontWeight: 600,
-                  }}
-                >
-                  €{option.standardPrice.toLocaleString()}
-                </Typography>
-              </Box>
-            </Stack>
-            <Typography
-              variant="body2"
-              color="text.secondary"
-              sx={{
-                lineHeight: 1.5,
-                mb: 1,
-                wordBreak: "break-word",
-                hyphens: "auto",
-              }}
-            >
-              {option.description}
-            </Typography>
-            {option.benefits && (
-              <Stack
-                direction="row"
-                spacing={0.5}
-                flexWrap="wrap"
-                useFlexGap
-                mb={1}
-              >
-                {option.benefits.map((benefit) => (
-                  <Chip
-                    key={benefit}
-                    label={benefit}
-                    size="small"
-                    sx={{
-                      bgcolor: "primary.light",
-                      color: "primary.contrastText",
-                      fontSize: { xs: "0.65rem", sm: "0.75rem" },
-                      height: { xs: 20, sm: 24 },
-                    }}
-                  />
-                ))}
-              </Stack>
-            )}
-            {option.includes && (
-              <Box mb={1}>
-                <Typography
-                  variant="subtitle2"
-                  gutterBottom
-                  fontWeight={600}
-                  sx={{ fontSize: { xs: "0.8rem", sm: "0.875rem" } }}
-                >
-                  Includes:
-                </Typography>
-                <Stack direction="row" spacing={0.5} flexWrap="wrap" useFlexGap>
-                  {option.includes
-                    .slice(0, expanded ? option.includes.length : 3)
-                    .map((item) => (
-                      <Chip
-                        key={item}
-                        label={item}
-                        size="small"
-                        sx={{
-                          bgcolor: "grey.100",
-                          fontSize: { xs: "0.65rem", sm: "0.75rem" },
-                          height: { xs: 20, sm: 24 },
-                        }}
-                      />
-                    ))}
-                  {option.includes.length > 3 && !expanded && (
-                    <Button
-                      size="small"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        setExpanded(true);
-                      }}
-                      sx={{
-                        minWidth: "auto",
-                        p: 0.5,
-                        fontSize: { xs: "0.65rem", sm: "0.75rem" },
-                      }}
-                    >
-                      +{option.includes.length - 3} more
-                    </Button>
-                  )}
-                </Stack>
-              </Box>
-            )}
-            {option.features && (
-              <Box>
-                <Button
-                  size="small"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    setExpanded(!expanded);
-                  }}
-                  endIcon={
-                    expanded ? (
-                      <ChevronUp size={16} />
-                    ) : (
-                      <ChevronDown size={16} />
-                    )
-                  }
-                  sx={{
-                    mb: 1,
-                    p: 0,
-                    justifyContent: "flex-start",
-                    fontSize: { xs: "0.75rem", sm: "0.875rem" },
-                  }}
-                >
-                  {expanded ? "Hide" : "Show"} Features (
-                  {option.features.length})
-                </Button>
-
-                <Collapse in={expanded}>
-                  <List
-                    dense
-                    sx={{
-                      py: 0,
-                      maxHeight: { xs: 200, sm: "none" },
-                      overflowY: { xs: "auto", sm: "visible" },
-                    }}
-                  >
-                    {option.features.map((feature, idx) => (
-                      <ListItem
-                        key={idx}
-                        disableGutters
-                        sx={{
-                          py: 0.25,
-                          px: 0,
-                        }}
-                      >
-                        <ListItemIcon sx={{ minWidth: 20 }}>
-                          <Check size={12} color={theme.palette.success.main} />
-                        </ListItemIcon>
-                        <ListItemText
-                          primary={feature}
-                          primaryTypographyProps={{
-                            variant: "body2",
-                            sx: {
-                              lineHeight: 1.4,
-                              fontSize: { xs: "0.75rem", sm: "0.875rem" },
-                            },
-                          }}
-                        />
-                      </ListItem>
-                    ))}
-                  </List>
-                </Collapse>
-              </Box>
-            )}
-            {option.savings && (
-              <Typography
-                variant="body2"
-                sx={{
-                  color: "success.main",
-                  fontWeight: 600,
-                  display: "flex",
-                  alignItems: "center",
-                  gap: 0.5,
-                  fontSize: { xs: "0.75rem", sm: "0.875rem" },
-                }}
-              >
-                <LocalOffer size={14} />
-                Save €{option.savings.toLocaleString()} vs individual
-              </Typography>
-            )}
-            {option.note && (
-              <Alert
-                severity="warning"
-                sx={{
-                  mt: 1,
-                  borderRadius: 2,
-                  fontSize: { xs: "0.75rem", sm: "0.875rem" },
-                  "& .MuiAlert-message": {
-                    py: 0,
-                  },
-                }}
-              >
-                {option.note}
-              </Alert>
-            )}
-          </Stack>
-        </Stack>
-      </Stack>
-    </SelectableCard>
-  );
-};
-
-TrinityPackageCard.propTypes = {
-  option: PropTypes.object.isRequired,
-  selected: PropTypes.bool,
-  onClick: PropTypes.func,
-};
-const TrinityPackageSelection = ({
-  selectedTrinity,
-  setSelectedTrinity,
-  nextStep,
-  prevStep,
-  currentServiceIndex,
-  totalServices,
-  isPartOfPlatform = false,
-}) => (
-  <Fade in timeout={500}>
-    <Box sx={{ px: { xs: 1, sm: 0 } }}>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label={
-            isPartOfPlatform
-              ? `Service ${
-                  currentServiceIndex + 1
-                } of ${totalServices}: Platform Systems`
-              : `Service ${
-                  currentServiceIndex + 1
-                } of ${totalServices}: Trinity Systems`
-          }
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          Choose Your Trinity Solution
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 700, fontSize: "1.1rem", lineHeight: 1.7, mb: 2 }}
-        >
-          Select individual AI systems or complete packages. Limited-time beta
-          pricing available!
-        </Typography>
-        <Chip
-          label="🔥 BETA PRICING: Save up to 87%"
-          color="error"
-          variant="filled"
-          sx={{
-            bgcolor: "error.main",
-            color: "white",
-            fontWeight: 700,
-            fontSize: "0.9rem",
-            px: 3,
-            py: 1,
-            animation: "pulse 2s infinite",
-            "@keyframes pulse": {
-              "0%": { opacity: 1 },
-              "50%": { opacity: 0.8 },
-              "100%": { opacity: 1 },
-            },
-          }}
-        />
-      </Stack>
-
-      <Grid container spacing={{ xs: 2, sm: 3, md: 4 }} sx={{ mb: 6 }}>
-        {TRINITY_OPTIONS.map((option, index) => (
-          <Grid
-            item
-            xs={12}
-            md={option.type === "package" ? 12 : 6}
-            key={option.id}
-          >
-            <Grow in timeout={300 + index * 100}>
-              <div>
-                <TrinityPackageCard
-                  option={option}
-                  selected={selectedTrinity === option.id}
-                  onClick={() => setSelectedTrinity(option.id)}
-                />
-              </div>
-            </Grow>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Stack
-        direction={{ xs: "column", sm: "row" }}
-        spacing={2}
-        justifyContent="center"
-        alignItems="center"
-      >
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{
-            minWidth: { xs: "100%", sm: 120 },
-            maxWidth: { xs: 300, sm: "none" },
-          }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={!selectedTrinity}
-          endIcon={<ChevronRight size={20} />}
-          sx={{
-            minWidth: { xs: "100%", sm: 160 },
-            maxWidth: { xs: 300, sm: "none" },
-          }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-TrinityPackageSelection.propTypes = {
-  selectedTrinity: PropTypes.string,
-  setSelectedTrinity: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-  currentServiceIndex: PropTypes.number,
-  totalServices: PropTypes.number,
-  isPartOfPlatform: PropTypes.bool,
-};
-const PlatformTierSelection = ({
-  selectedTier,
-  setSelectedTier,
-  nextStep,
-  prevStep,
-  currentServiceIndex,
-  totalServices,
-}) => (
-  <Fade in timeout={500}>
-    <Box>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label={`Service ${
-            currentServiceIndex + 1
-          } of ${totalServices}: Platform Tier`}
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          Select Your Platform Tier
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 700, fontSize: "1.1rem", lineHeight: 1.7 }}
-        >
-          Choose the functionality level that best matches your business
-          requirements and growth plans.
-        </Typography>
-      </Stack>
-
-      <Grid container spacing={4} sx={{ mb: 6 }}>
-        {PLATFORM_TIERS.map((tier, index) => (
-          <Grid item xs={12} md={4} key={tier.id}>
-            <Grow in timeout={300 + index * 150}>
-              <div>
-                <SelectableCard
-                  selected={selectedTier === tier.id}
-                  onClick={() => setSelectedTier(tier.id)}
-                  sx={{ height: "100%" }}
-                >
-                  {tier.popular && (
-                    <Chip
-                      label="Most Popular"
-                      color="primary"
-                      size="small"
-                      sx={{
-                        position: "absolute",
-                        top: -10,
-                        left: "50%",
-                        transform: "translateX(-50%)",
-                        zIndex: 1,
-                        bgcolor: "primary.main",
-                        color: "white",
-                        fontWeight: 600,
-                      }}
-                    />
-                  )}
-
-                  <Stack spacing={2} height="100%">
-                    <Box textAlign="center" pt={tier.popular ? 2 : 0}>
-                      <Typography
-                        variant="h3"
-                        sx={{ mb: 1, fontSize: "2.5rem" }}
-                      >
-                        {tier.icon}
-                      </Typography>
-                      <Typography
-                        variant="h5"
-                        component="h3"
-                        gutterBottom
-                        fontWeight={700}
-                      >
-                        {tier.name}
-                      </Typography>
-                      <Typography
-                        variant="body2"
-                        color="text.secondary"
-                        sx={{ mb: 2, lineHeight: 1.6 }}
-                      >
-                        {tier.description}
-                      </Typography>
-                      <Box sx={{ mb: 2 }}>
-                        <Typography
-                          variant="h4"
-                          component="p"
-                          color="primary"
-                          sx={{ fontWeight: 800, mb: 0.5 }}
-                        >
-                          €{tier.minPrice.toLocaleString()} - €
-                          {tier.maxPrice.toLocaleString()}
-                        </Typography>
-                        <Chip
-                          label={tier.deliveryTime}
-                          size="small"
-                          icon={<Schedule size={14} />}
-                          sx={{ bgcolor: "grey.100", fontSize: "0.75rem" }}
-                        />
-                      </Box>
-                    </Box>
-
-                    <List dense sx={{ p: 0, flexGrow: 1 }}>
-                      {tier.features.map((feature, idx) => (
-                        <ListItem key={idx} disableGutters sx={{ py: 0.5 }}>
-                          <ListItemIcon sx={{ minWidth: 32 }}>
-                            <Box
-                              sx={{
-                                width: 20,
-                                height: 20,
-                                borderRadius: "50%",
-                                bgcolor: "success.main",
-                                display: "flex",
-                                alignItems: "center",
-                                justifyContent: "center",
-                              }}
-                            >
-                              <Check size={12} color="white" />
-                            </Box>
-                          </ListItemIcon>
-                          <ListItemText
-                            primary={feature}
-                            primaryTypographyProps={{
-                              variant: "body2",
-                              sx: { lineHeight: 1.6 },
-                            }}
-                          />
-                        </ListItem>
-                      ))}
-                    </List>
-                  </Stack>
-                </SelectableCard>
-              </div>
-            </Grow>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Stack direction="row" spacing={2} justifyContent="center">
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{ minWidth: 120 }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={!selectedTier}
-          endIcon={<ChevronRight size={20} />}
-          sx={{ minWidth: 160 }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-PlatformTierSelection.propTypes = {
-  selectedTier: PropTypes.string,
-  setSelectedTier: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-  currentServiceIndex: PropTypes.number,
-  totalServices: PropTypes.number,
-};
-const StoreTypeSelection = ({
-  hasPhysicalStore,
-  setHasPhysicalStore,
-  nextStep,
-  prevStep,
-}) => (
-  <Fade in timeout={500}>
-    <Box>
-      <Stack spacing={3} alignItems="center" textAlign="center" mb={6}>
-        <Chip
-          label="Business Type Configuration"
-          color="primary"
-          variant="outlined"
-          sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-        />
-        <Typography variant="h2" gutterBottom>
-          Do you have a physical storefront?
-        </Typography>
-        <Typography
-          variant="body1"
-          color="text.secondary"
-          sx={{ maxWidth: 600, fontSize: "1.1rem", lineHeight: 1.7 }}
-        >
-          This affects GARO system setup requirements and pricing. Physical
-          stores require additional Square POS integration.
-        </Typography>
-      </Stack>
-
-      <Grid container spacing={4} sx={{ mb: 6, justifyContent: "center" }}>
-        <Grid item xs={12} md={5}>
-          <Grow in timeout={400}>
-            <div>
-              <SelectableCard
-                selected={hasPhysicalStore === false}
-                onClick={() => setHasPhysicalStore(false)}
-                sx={{ textAlign: "center" }}
-              >
-                <Box
-                  sx={{
-                    width: 80,
-                    height: 80,
-                    borderRadius: "50%",
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    mb: 3,
-                    mx: "auto",
-                    background: "linear-gradient(135deg, #6366F1, #8B5CF6)",
-                    boxShadow: "0 8px 25px rgba(99, 102, 241, 0.3)",
-                  }}
-                >
-                  <Wifi size={36} color="white" />
-                </Box>
-                <Typography
-                  variant="h5"
-                  component="h3"
-                  gutterBottom
-                  fontWeight={700}
-                >
-                  E-commerce Only
-                </Typography>
-                <Typography
-                  variant="body1"
-                  color="text.secondary"
-                  sx={{ mb: 3, lineHeight: 1.6 }}
-                >
-                  Online business without a physical location. Perfect for
-                  digital-first companies.
-                </Typography>
-                <Chip
-                  label="Standard API setup included"
-                  color="success"
-                  variant="filled"
-                  size="small"
-                  sx={{ fontWeight: 600 }}
-                />
-              </SelectableCard>
-            </div>
-          </Grow>
-        </Grid>
-        <Grid item xs={12} md={5}>
-          <Grow in timeout={600}>
-            <div>
-              <SelectableCard
-                selected={hasPhysicalStore === true}
-                onClick={() => setHasPhysicalStore(true)}
-                sx={{ textAlign: "center" }}
-              >
-                <Box
-                  sx={{
-                    width: 80,
-                    height: 80,
-                    borderRadius: "50%",
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    mb: 3,
-                    mx: "auto",
-                    background: "linear-gradient(135deg, #059669, #10B981)",
-                    boxShadow: "0 8px 25px rgba(5, 150, 105, 0.3)",
-                  }}
-                >
-                  <Building size={36} color="white" />
-                </Box>
-                <Typography
-                  variant="h5"
-                  component="h3"
-                  gutterBottom
-                  fontWeight={700}
-                >
-                  Physical Storefront
-                </Typography>
-                <Typography
-                  variant="body1"
-                  color="text.secondary"
-                  sx={{ mb: 3, lineHeight: 1.6 }}
-                >
-                  Brick-and-mortar location with inventory management needs and
-                  POS integration.
-                </Typography>
-                <Chip
-                  label="Includes €1,600 Square setup & training"
-                  color="warning"
-                  variant="filled"
-                  size="small"
-                  sx={{ fontWeight: 600 }}
-                />
-              </SelectableCard>
-            </div>
-          </Grow>
-        </Grid>
-      </Grid>
-
-      <Stack direction="row" spacing={2} justifyContent="center">
-        <Button
-          variant="outlined"
-          size="large"
-          onClick={prevStep}
-          startIcon={<ChevronLeft size={20} />}
-          sx={{ minWidth: 120 }}
-        >
-          Back
-        </Button>
-        <Button
-          variant="contained"
-          size="large"
-          onClick={nextStep}
-          disabled={hasPhysicalStore === null}
-          endIcon={<ChevronRight size={20} />}
-          sx={{ minWidth: 160 }}
-        >
-          Continue
-        </Button>
-      </Stack>
-    </Box>
-  </Fade>
-);
-
-StoreTypeSelection.propTypes = {
-  hasPhysicalStore: PropTypes.bool,
-  setHasPhysicalStore: PropTypes.func.isRequired,
-  nextStep: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-};
-
-const FinalSummary = ({
-  selectedServices,
-  selectedIndustry,
-  serviceSelections,
-  hasPhysicalStore,
-  name,
-  setName,
-  email,
-  setEmail,
-  additionalNotes,
-  setAdditionalNotes,
-  keywords,
-  setKeywords,
-  selectedSystems,
-  setSelectedSystems,
-  selectedDashboards,
-  setSelectedDashboards,
-  prevStep,
-  handleCheckout,
-  handleConsultationCheckout,
-  isProcessing,
-  isProcessingConsult,
-  calculateTotal,
-}) => {
-  const total = calculateTotal();
-
-  const getServiceName = (serviceId) => {
-    const service = SERVICES.find((s) => s.id === serviceId);
-    return service?.name || "Unknown Service";
-  };
-
-  const getServiceSelectionDetails = (serviceId) => {
-    const selection = serviceSelections[serviceId];
-    let details = [];
-    let price = 0;
-
-    if (serviceId === "brand-identity" && selection?.package) {
-      const pkg = BRAND_PACKAGES.find((p) => p.id === selection.package);
-      if (pkg) {
-        details.push(pkg.name);
-        price += pkg.price;
-      }
-    }
-
-    if (serviceId === "ad-campaigns" && selection?.campaign) {
-      const campaign = CAMPAIGN_PACKAGES.find(
-        (p) => p.id === selection.campaign
-      );
-      if (campaign) {
-        details.push(campaign.name);
-        price += campaign.price;
-      }
-    }
-
-    if (
-      (serviceId === "trinity-systems" || serviceId === "complete-platform") &&
-      selection?.trinity
-    ) {
-      const trinity = TRINITY_OPTIONS.find((p) => p.id === selection.trinity);
-      if (trinity) {
-        details.push(trinity.name);
-        price += trinity.betaPrice;
-      }
-    }
-
-    if (serviceId === "complete-platform" && selection?.tier) {
-      const tier = PLATFORM_TIERS.find((t) => t.id === selection.tier);
-      if (tier) {
-        details.push(`${tier.name} Platform`);
-        price += (tier.minPrice + tier.maxPrice) / 2;
-      }
-    }
-
-    return { details, price };
-  };
-
-  const industryInfo = industries.find((i) => i.id === selectedIndustry);
-
-  return (
-    <Fade in timeout={500}>
-      <Grid container spacing={4}>
-        <Grid item xs={12} lg={8}>
-          <Stack spacing={4}>
-            <Box>
-              <Stack spacing={3} alignItems="center" textAlign="center" mb={4}>
-                <Chip
-                  label="Final Step: Review & Purchase"
-                  color="primary"
-                  variant="outlined"
-                  sx={{ px: 3, py: 1, fontSize: "0.9rem", fontWeight: 600 }}
-                />
-                <Typography variant="h2" gutterBottom>
-                  Confirm Your Configuration
-                </Typography>
-                <Typography
-                  variant="body1"
-                  color="text.secondary"
-                  sx={{ maxWidth: 600, fontSize: "1.1rem", lineHeight: 1.7 }}
-                >
-                  Review your selections and provide your details to proceed
-                  with your custom solution.
-                </Typography>
-              </Stack>
-            </Box>
-
-            <Paper
-              elevation={0}
-              sx={{ p: 4, border: "1px solid", borderColor: "grey.200" }}
-            >
-              <Typography variant="h5" gutterBottom fontWeight={700} mb={3}>
-                Your Details
-              </Typography>
-              <Grid container spacing={3}>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    fullWidth
-                    label="Your Name *"
-                    value={name}
-                    onChange={(e) => setName(e.target.value)}
-                    required
-                    variant="outlined"
-                  />
-                </Grid>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    fullWidth
-                    label="Your Email *"
-                    type="email"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    required
-                    variant="outlined"
-                  />
-                </Grid>
-              </Grid>
-            </Paper>
-
-            <Paper
-              elevation={0}
-              sx={{ p: 4, border: "1px solid", borderColor: "grey.200" }}
-            >
-              <Typography variant="h5" gutterBottom fontWeight={700} mb={3}>
-                Optional Customization
-              </Typography>
-              <Grid container spacing={3}>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    fullWidth
-                    label="Keywords (comma-separated)"
-                    value={keywords}
-                    onChange={(e) => setKeywords(e.target.value)}
-                    placeholder="marketing, sales, growth"
-                  />
-                </Grid>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    fullWidth
-                    label="Preferred Systems"
-                    value={selectedSystems}
-                    onChange={(e) => setSelectedSystems(e.target.value)}
-                    placeholder="CRM, ERP, Analytics"
-                  />
-                </Grid>
-                <Grid item xs={12} sm={6}>
-                  <TextField
-                    fullWidth
-                    label="Dashboard Preferences"
-                    value={selectedDashboards}
-                    onChange={(e) => setSelectedDashboards(e.target.value)}
-                    placeholder="Sales, Marketing, Finance"
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    fullWidth
-                    multiline
-                    rows={4}
-                    label="Additional Notes"
-                    value={additionalNotes}
-                    onChange={(e) => setAdditionalNotes(e.target.value)}
-                    placeholder="Tell us about any specific requirements or questions you have..."
-                  />
-                </Grid>
-              </Grid>
-            </Paper>
-          </Stack>
-        </Grid>
-
-        <Grid item xs={12} lg={4}>
-          <Paper
-            elevation={0}
-            sx={{
-              position: "sticky",
-              top: "24px",
-              border: "1px solid",
-              borderColor: "grey.200",
-              borderRadius: 3,
-            }}
-          >
-            <Box sx={{ p: 3 }}>
-              <Typography variant="h5" gutterBottom fontWeight={700}>
-                Order Summary
-              </Typography>
-              <Divider sx={{ my: 2 }} />
-
-              <Stack spacing={3}>
-                <Box>
-                  <Typography
-                    variant="h6"
-                    gutterBottom
-                    fontWeight={600}
-                    color="primary"
-                  >
-                    Selected Services ({selectedServices.length}):
-                  </Typography>
-                  {selectedServices.map((serviceId) => {
-                    const { details, price } =
-                      getServiceSelectionDetails(serviceId);
-                    return (
-                      <Paper
-                        key={serviceId}
-                        elevation={0}
-                        sx={{
-                          mb: 2,
-                          p: 2,
-                          bgcolor: "grey.50",
-                          borderRadius: 2,
-                          border: "1px solid",
-                          borderColor: "grey.200",
-                        }}
-                      >
-                        <Typography variant="subtitle1" fontWeight={700}>
-                          {getServiceName(serviceId)}
-                        </Typography>
-                        {details.map((detail, idx) => (
-                          <Typography
-                            key={idx}
-                            variant="body2"
-                            color="text.secondary"
-                            sx={{ ml: 1 }}
-                          >
-                            • {detail}
-                          </Typography>
-                        ))}
-                        <Typography
-                          variant="h6"
-                          color="primary"
-                          sx={{ mt: 1, fontWeight: 700 }}
-                        >
-                          €{price.toLocaleString()}
-                        </Typography>
-                      </Paper>
-                    );
-                  })}
-                </Box>
-
-                {industryInfo && (
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography variant="body1" fontWeight={600}>
-                      Industry:
-                    </Typography>
-                    <Chip
-                      label={industryInfo.name}
-                      icon={<Business size={16} />}
-                      size="small"
-                      variant="outlined"
-                    />
-                  </Stack>
-                )}
-
-                {hasPhysicalStore !== null && (
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography variant="body1" fontWeight={600}>
-                      Store Type:
-                    </Typography>
-                    <Chip
-                      label={
-                        hasPhysicalStore ? "Physical Store" : "E-commerce Only"
-                      }
-                      icon={
-                        hasPhysicalStore ? (
-                          <Building size={16} />
-                        ) : (
-                          <Globe size={16} />
-                        )
-                      }
-                      size="small"
-                      variant="outlined"
-                    />
-                  </Stack>
-                )}
-
-                {hasPhysicalStore && (
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography variant="body1" fontWeight={600}>
-                      Setup Fee:
-                    </Typography>
-                    <Typography variant="h6" fontWeight={700}>
-                      €1,600
-                    </Typography>
-                  </Stack>
-                )}
-              </Stack>
-
-              <Box mt={4} pt={3} borderTop={1} borderColor="grey.300">
-                <Stack
-                  direction="row"
-                  justifyContent="space-between"
-                  alignItems="center"
-                  mb={2}
-                >
-                  <Typography variant="body1" fontWeight={700}>
-                    Total Price:
-                  </Typography>
-                  <Typography
-                    variant="body1"
-                    fontWeight={800}
-                    strikeThrough
-                    // sx={{
-                    //   background:
-                    //     "linear-gradient(135deg, #6366F1 0%, #8B5CF6 50%, #EC4899 100%)",
-                    //   WebkitBackgroundClip: "text",
-                    //   WebkitTextFillColor: "transparent",
-                    //   backgroundClip: "text",
-            
-                    // }}
-                    sx={{
-                      color: "primary.main",
-                      textDecoration: "line-through",
-                    }}
-                    
-                  >
-                    €{total.toLocaleString()}
-                  </Typography>
-                </Stack>
-                <Stack
-                  direction="row"
-                  justifyContent="space-between"
-                  alignItems="center"
-                  mb={2}
-                >
-                  <Typography variant="h5" fontWeight={700}>
-                    Deposit Price:
-                  </Typography>
-                  <Typography
-                    variant="h4"
-                    fontWeight={800}
-                    sx={{
-                      background:
-                        "linear-gradient(135deg, #6366F1 0%, #8B5CF6 50%, #EC4899 100%)",
-                      WebkitBackgroundClip: "text",
-                      WebkitTextFillColor: "transparent",
-                      backgroundClip: "text",
-                    }}
-                  >
-                    €{applyDiscount(total).toLocaleString()}
-                  </Typography>
-                </Stack>
-                <Typography
-                  variant="caption"
-                  color="success.main"
-                  display="block"
-                  textAlign="right"
-                  mb={2}
-                >
-                  {total < 1000
-                    ? "50% deposit applied"
-                    : "20% deposit applied"}
-                </Typography>
-                <Stack spacing={2}>
-                  <Button
-                    variant="contained"
-                    fullWidth
-                    size="large"
-                    onClick={handleCheckout}
-                    disabled={isProcessing || !name || !email}
-                    startIcon={
-                      isProcessing ? (
-                        <CircularProgress size={20} color="inherit" />
-                      ) : (
-                        <Rocket size={20} />
-                      )
-                    }
-                    sx={{
-                      py: 2,
-                      fontSize: "1rem",
-                      fontWeight: 700,
-                      background:
-                        "linear-gradient(135deg, #6366F1 0%, #8B5CF6 100%)",
-                      "&:hover": {
-                        background:
-                          "linear-gradient(135deg, #4F46E5 0%, #7C3AED 100%)",
-                        transform: "translateY(-1px)",
-                        boxShadow: "0 8px 25px rgba(99, 102, 241, 0.4)",
-                        color: "#FFFFFF",
-                      },
-                    }}
-                  >
-                    {isProcessing ? "Processing..." : "Proceed to Checkout"}
-                  </Button>
-
-                  <Button
-                    variant="outlined"
-                    fullWidth
-                    onClick={handleConsultationCheckout}
-                    disabled={isProcessingConsult || !name || !email}
-                    sx={{
-                      borderColor: "primary.main",
-                      color: "primary.main",
-                      "&:hover": {
-                        borderColor: "primary.dark",
-                        backgroundColor: "primary.main",
-                        color: "white",
-                      },
-                    }}
-                    startIcon={
-                      isProcessingConsult ? (
-                        <CircularProgress size={20} color="inherit" />
-                      ) : null
-                    }
-                  >
-                    {isProcessingConsult
-                      ? "Processing..."
-                      : "Book a Consultation (€83)"}
-                  </Button>
-
-                  <Button
-                    variant="text"
-                    fullWidth
-                    onClick={prevStep}
-                    startIcon={<ChevronLeft size={20} />}
-                    sx={{
-                      py: 1.5,
-                      color: "text.secondary",
-                      fontWeight: 600,
-                    }}
-                  >
-                    Back to Edit
-                  </Button>
-                </Stack>
-              </Box>
-            </Box>
-          </Paper>
-        </Grid>
-      </Grid>
-    </Fade>
-  );
-};
-
-FinalSummary.propTypes = {
-  selectedServices: PropTypes.array.isRequired,
-  selectedIndustry: PropTypes.string,
-  serviceSelections: PropTypes.object.isRequired,
-  hasPhysicalStore: PropTypes.bool,
-  name: PropTypes.string.isRequired,
-  setName: PropTypes.func.isRequired,
-  email: PropTypes.string.isRequired,
-  setEmail: PropTypes.func.isRequired,
-  additionalNotes: PropTypes.string,
-  setAdditionalNotes: PropTypes.func.isRequired,
-  keywords: PropTypes.string,
-  setKeywords: PropTypes.func.isRequired,
-  selectedSystems: PropTypes.string,
-  setSelectedSystems: PropTypes.func.isRequired,
-  selectedDashboards: PropTypes.string,
-  setSelectedDashboards: PropTypes.func.isRequired,
-  prevStep: PropTypes.func.isRequired,
-  handleCheckout: PropTypes.func.isRequired,
-  handleConsultationCheckout: PropTypes.func.isRequired,
-  isProcessing: PropTypes.bool.isRequired,
-  isProcessingConsult: PropTypes.bool.isRequired,
-  calculateTotal: PropTypes.func.isRequired,
-};
 
 const MultiStepForm = () => {
   const [currentStep, setCurrentStep] = useState(1);
@@ -3074,11 +516,11 @@ const MultiStepForm = () => {
           sx={{ py: { xs: 18, md: 20 } }}
           ref={wizardRef}
         >
-          <Fade in timeout={1000}>
+          {/* <Fade in={true} timeout={400}> */}
             <Box textAlign="center" mb={{ xs: 6, md: 10 }}>
               <Typography variant="h1" gutterBottom>
-                Your Partner in Accelerating the{" "}
-                <Box component="span" sx={gradientText}>
+                Your Partner in Accelerating the
+                <Box component="span" sx={GradientText}>
                   Digital Space
                 </Box>
               </Typography>
@@ -3122,9 +564,9 @@ const MultiStepForm = () => {
                 Trusted by 500+ businesses worldwide
               </Typography>
             </Box>
-          </Fade>
+          {/* </Fade> */}
 
-          <Fade in timeout={1200}>
+          {/* <Fade in timeout={500}> */}
             <Box mb={{ xs: 6, md: 8 }}>
               <Stepper
                 activeStep={currentStep - 1}
@@ -3145,10 +587,10 @@ const MultiStepForm = () => {
                   },
                 }}
               >
-                {steps.map((label, index) => (
+                {steps.map((label) => (
                   <Step key={label}>
                     <StepLabel
-                      StepIconComponent={({ active, completed }) => (
+                      StepIconComponent={({ active, completed, icon }) => (
                         <Box
                           sx={{
                             width: 40,
@@ -3160,15 +602,15 @@ const MultiStepForm = () => {
                             bgcolor: completed
                               ? "success.main"
                               : active
-                              ? "primary.main"
-                              : "grey.300",
+                                ? "primary.main"
+                                : "grey.300",
                             color: "white",
                             fontWeight: 700,
                             fontSize: "1rem",
                             transition: "all 0.3s ease",
                           }}
                         >
-                          {completed ? <Check size={20} /> : index + 1}
+                          {completed ? <Check size={20} /> : icon}
                         </Box>
                       )}
                     >
@@ -3178,7 +620,7 @@ const MultiStepForm = () => {
                 ))}
               </Stepper>
             </Box>
-          </Fade>
+          {/* </Fade> */}
           {renderStepContent()}
         </Container>
         <Snackbar
@@ -3198,8 +640,8 @@ const MultiStepForm = () => {
               showToast && showToast.toLowerCase().includes("error")
                 ? "error"
                 : showToast && showToast.toLowerCase().includes("success")
-                ? "success"
-                : "info"
+                  ? "success"
+                  : "info"
             }
             sx={{
               width: "100%",
