+++ src/components/Services/CostCalculation/data/costCalculationData.js
@@ -1,0 +1,404 @@
+import {
+  Rocket,
+  Palette,
+  Megaphone,
+  Settings,
+
+} from "lucide-react";
+
+export const SERVICES = [
+  {
+    id: "brand-identity",
+    name: "Brand Identity Package",
+    icon: Palette,
+    description:
+      "Professional brand identity design and strategy that sets you apart",
+    color: "#E11D48",
+    gradient: "linear-gradient(135deg, #E11D48 0%, #F97316 100%)",
+    benefits: ["Logo Design", "Brand Guidelines", "Marketing Materials"],
+  },
+  {
+    id: "ad-campaigns",
+    name: "Digital Ad Campaigns",
+    icon: Megaphone,
+    description: "Targeted digital marketing campaigns that drive real results",
+    color: "#059669",
+    gradient: "linear-gradient(135deg, #059669 0%, #10B981 100%)",
+    benefits: ["Google Ads", "Meta Ads", "Performance Analytics"],
+  },
+  {
+    id: "trinity-systems",
+    name: "Trinity Business Systems",
+    icon: Settings,
+    description: "AI-powered business intelligence platforms for growth",
+    color: "#7C3AED",
+    gradient: "linear-gradient(135deg, #7C3AED 0%, #6366F1 100%)",
+    benefits: ["AI Optimization", "Predictive Analytics", "Automation"],
+  },
+  {
+    id: "complete-platform",
+    name: "Complete Digital Platform",
+    icon: Rocket,
+    description: "Full-service website and business systems solution",
+    color: "#DC2626",
+    gradient: "linear-gradient(135deg, #DC2626 0%, #EA580C 100%)",
+    benefits: ["Custom Website", "E-commerce", "Full Integration"],
+  },
+];
+
+export const BRAND_PACKAGES = [
+  {
+    id: "starter",
+    name: "Starter Brand Identity",
+    price: 500,
+    popular: false,
+    description: "Perfect for new businesses starting their brand journey",
+    features: [
+      "Logo design concept (up to 2 proposals)",
+      "Colour scheme development",
+      "Typography selection",
+      "Business card design",
+      "Letterhead or invoice design",
+      "Brand questionnaire & style scope",
+      "1 revision round",
+    ],
+    deliveryTime: "5-7 business days",
+  },
+  {
+    id: "premium",
+    name: "Premium Brand Identity",
+    price: 750,
+    popular: true,
+    description: "Most popular choice for established businesses",
+    features: [
+      "Logo design concept (up to 2 proposals)",
+      "Comprehensive colour scheme",
+      "Typography system",
+      "Business card design",
+      "Letterhead or invoice design",
+      "Brand questionnaire & style scope",
+      "Brand dos and don'ts guide",
+      "Up to 3 marketing materials",
+      "2 revision rounds",
+    ],
+    deliveryTime: "7-10 business days",
+  },
+  {
+    id: "elite",
+    name: "Elite Brand Identity",
+    price: 1000,
+    popular: false,
+    description: "Complete brand system for premium businesses",
+    features: [
+      "Logo design concept (up to 2 proposals)",
+      "Full colour system",
+      "Typography hierarchy",
+      "Complete stationery suite",
+      "Brand questionnaire & style scope",
+      "Comprehensive brand guidelines",
+      "Up to 5 marketing materials",
+      "4 revision rounds",
+      "Brand personality overview",
+      "Printing & digital asset guides",
+      "1-hour consultation meeting",
+    ],
+    deliveryTime: "10-14 business days",
+  },
+];
+
+export const CAMPAIGN_PACKAGES = [
+  {
+    id: "starter",
+    name: "Starter Campaign",
+    price: 350,
+    description: "Perfect for small businesses starting digital marketing",
+    icon: "🚀",
+    features: [
+      "Basic customer segmentation",
+      "Competitor scan + keyword research (10-15 keywords)",
+      "1x Google Ads campaign setup",
+      "Up to 4 custom ad designs",
+      "Performance report after 2 months",
+    ],
+    deliveryTime: "3-5 business days",
+  },
+  {
+    id: "growth",
+    name: "Growth Campaign",
+    price: 500,
+    description: "Most popular for SMEs seeking wider reach",
+    featured: true,
+    icon: "📈",
+    features: [
+      "Advanced customer segmentation",
+      "Comprehensive competitor analysis",
+      "2x Google Ads campaigns with optimization",
+      "Up to 7 premium ad designs",
+      "Mid-campaign update + final report",
+      "CTR, conversions & ROI insights",
+    ],
+    deliveryTime: "5-7 business days",
+  },
+  {
+    id: "premium",
+    name: "Premium Campaign",
+    price: 750,
+    description: "Complete solution for market domination",
+    icon: "🏆",
+    features: [
+      "Deluxe segmentation + 2 buyer personas",
+      "In-depth competitor benchmarking",
+      "3x Google Ads campaigns with ongoing optimization",
+      "Up to 10 premium ad designs with copywriting",
+      "Detailed analytics dashboard",
+      "Strategic recommendations & next steps",
+    ],
+    deliveryTime: "7-10 business days",
+  },
+];
+
+export const TRINITY_OPTIONS = [
+  {
+    id: "expense",
+    type: "individual",
+    name: "Expense Manager",
+    icon: "💰",
+    description:
+      "AI-powered financial planning that ensures you never run out of cash",
+    betaPrice: 230,
+    standardPrice: 2074,
+    savings: 87,
+    features: [
+      "Predictive cash flow modeling",
+      "7 intelligent allocation rules",
+      "Emergency fund protection",
+      "Automatic priority reshuffling",
+      "Real-time expense tracking",
+    ],
+    benefits: [
+      "Avoid cash shortages",
+      "Optimize spending",
+      "Build emergency funds",
+    ],
+  },
+  {
+    id: "mcd",
+    type: "individual",
+    name: "MCD System",
+    icon: "📈",
+    description:
+      "Marketing Cost Displacement - automatically adjusts prices based on ad spend",
+    betaPrice: 200,
+    standardPrice: 1500,
+    savings: 87,
+    features: [
+      "Real-time price optimization",
+      "Marketing spend integration",
+      "Automatic profit protection",
+      "Platform cost tracking",
+      "Dynamic pricing algorithms",
+    ],
+    benefits: [
+      "Maintain profit margins",
+      "Optimize ad ROI",
+      "Automated pricing",
+    ],
+  },
+  {
+    id: "rcd",
+    type: "individual",
+    name: "RCD System",
+    icon: "🎯",
+    description:
+      "Returning Customer Discounts - creates viral loyalty networks",
+    betaPrice: 200,
+    standardPrice: 1500,
+    savings: 87,
+    features: [
+      "Automatic loyalty tracking",
+      "Viral referral networks",
+      "Personalized discount vectors",
+      "Lifetime value optimization",
+      "Customer journey mapping",
+    ],
+    benefits: ["Increase customer retention", "Viral growth", "Higher LTV"],
+  },
+  {
+    id: "garo",
+    type: "individual",
+    name: "GARO System",
+    icon: "🧬",
+    description:
+      "Genetic Algorithm Restocking Optimizer - evolves perfect inventory decisions",
+    betaPrice: 2230,
+    standardPrice: 2074,
+    savings: 87,
+    features: [
+      "500+ generation evolution",
+      "Multi-objective fitness scoring",
+      "70% stockout reduction",
+      "Predictive demand modeling",
+      "Automated reorder points",
+    ],
+    benefits: ["Reduce stockouts", "Optimize inventory", "Predict demand"],
+    note: "Physical stores: includes €1,600 for Square setup & training",
+  },
+  {
+    id: "aed",
+    type: "individual",
+    name: "AED System",
+    icon: "🚀",
+    description:
+      "Advertising Efficiency Dashboard - unifies and optimizes all ad platforms",
+    betaPrice: 230,
+    standardPrice: 1500,
+    savings: 87,
+    features: [
+      "4+ platform integration",
+      "Real-time budget reallocation",
+      "Machine learning optimization",
+      "30% cost reduction average",
+      "Cross-platform analytics",
+    ],
+    benefits: ["Unified ad management", "Reduce ad costs", "ML optimization"],
+  },
+  {
+    id: "trinity-core",
+    type: "package",
+    name: "Trinity Core Package",
+    icon: "⚡",
+    description: "Essential business intelligence: Expense Manager + MCD + RCD",
+    betaPrice: 690,
+    standardPrice: 5187,
+    includes: ["💰 Expense Manager", "📈 MCD System", "🎯 RCD System"],
+    savings: 1728,
+    baseRecommended: true,
+    benefits: [
+      "Complete financial control",
+      "Automated pricing",
+      "Customer loyalty",
+    ],
+  },
+  {
+    id: "trinity-plus",
+    type: "package",
+    name: "Trinity Plus Package",
+    icon: "🌟",
+    description: "Complete suite: All 5 systems working in perfect harmony",
+    betaPrice: 2994,
+    standardPrice: 10489,
+    includes: [
+      "💰 Expense Manager",
+      "📈 MCD System",
+      "🎯 RCD System",
+      "🧬 GARO System",
+      "🚀 AED System",
+    ],
+    savings: 4724,
+    bestValue: true,
+    benefits: ["Complete automation", "Maximum savings", "Full integration"],
+    note: "GARO requires €1,600 for physical stores",
+  },
+];
+
+export const PLATFORM_TIERS = [
+  {
+    id: "foundation",
+    name: "Foundation",
+    minPrice: 600,
+    maxPrice: 1000,
+    description: "Perfect starting point for your digital presence",
+    icon: "🏗️",
+    features: [
+      "Professional responsive website",
+      "Mobile-first design",
+      "Basic SEO optimization",
+      "Contact forms & lead capture",
+      "Social media integration",
+      "SSL certificate & security",
+    ],
+    deliveryTime: "7-10 business days",
+  },
+  {
+    id: "full",
+    name: "Full System",
+    minPrice: 1200,
+    maxPrice: 2000,
+    description: "Complete business solution with advanced features",
+    icon: "🎯",
+    popular: true,
+    features: [
+      "Everything in Foundation",
+      "Complete ordering system",
+      "Customer database & CRM",
+      "Email automation workflows",
+      "Payment processing integration",
+      "Analytics & reporting dashboard",
+      "Inventory management",
+    ],
+    deliveryTime: "14-21 business days",
+  },
+  {
+    id: "premium",
+    name: "Premium Enterprise",
+    minPrice: 1800,
+    maxPrice: 3000,
+    description: "Enterprise-grade solution for scaling businesses",
+    icon: "🚀",
+    features: [
+      "Everything in Full System",
+      "Advanced analytics & insights",
+      "Multi-location support",
+      "API access & integrations",
+      "AI-powered predictions",
+      "Priority support & training",
+      "Custom integrations",
+    ],
+    deliveryTime: "21-28 business days",
+  },
+];
+
+export const industries = [
+  {
+    id: "restaurant",
+    name: "Restaurant & Food",
+    icon: "🍽️",
+    examples: "Restaurants, Takeaways, Cafes, Food Delivery",
+    color: "#EF4444",
+  },
+  {
+    id: "retail",
+    name: "Retail & E-commerce",
+    icon: "🛍️",
+    examples: "Online Stores, Boutiques, Fashion, Electronics",
+    color: "#8B5CF6",
+  },
+  {
+    id: "services",
+    name: "Professional Services",
+    icon: "💼",
+    examples: "Consulting, Legal, Financial, Healthcare",
+    color: "#059669",
+  },
+  {
+    id: "hospitality",
+    name: "Hospitality & Events",
+    icon: "🏨",
+    examples: "Hotels, Venues, Tourism, Event Planning",
+    color: "#DC2626",
+  },
+  {
+    id: "property",
+    name: "Property & Real Estate",
+    icon: "🏠",
+    examples: "Agents, Property Management, Development",
+    color: "#0891B2",
+  },
+  {
+    id: "other",
+    name: "Other Industry",
+    icon: "🚀",
+    examples: "Tell us about your unique business",
+    color: "#6366F1",
+  },
+];
